the chatsep.store is something we need for the ssl, no reason to care about it, it is jus something thats handling the certificate as far as I understood, In the Controller I added two lines with System.setPropert.... that is to add Ssl basically and then the welcomesocket (which should be a serversocket) is initialized as seen, nothing else is modified
Question: Should the socket that s accepting be a serversocket as well or is it okay with the basic socket? Also, what do we do with the Socket private variable, I deleted from the constructor because it was initialized to itself, not sure if that is how it should be? If so put it back and sorry :D